Changes motivated by code review
